backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
celery
redis
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
paramiko
pysnmp 